rt - refactor profile page code
